Use React hooks to call browser APIs
