+++ package.json
@@ -25,7 +25,7 @@
     "postbuild": "dts-bundle-generator src/index.tsx -o build/index.d.ts --no-check --no-banner --external-imports react tinycolor2 date-fns",
     "prepublishOnly": "tsc && pnpm lint && pnpm build",
     "start": "rollup -c -w",
-    "test": "jest"
+    "test": "jest --silent"
   },
   "dependencies": {
     "date-fns": "^4.1.0"
@@ -54,6 +54,8 @@
     "@storybook/react": "^8.3.5",
     "@storybook/react-vite": "^8.3.5",
     "@storybook/theming": "^8.3.5",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/react": "^16.0.1",
     "@types/eslint__js": "^8.42.3",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.0",
@@ -65,6 +67,7 @@
     "eslint-plugin-react-hooks": "^4.6.2",
     "globals": "^15.10.0",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.4.47",
     "prettier": "^3.3.2",
     "prism-react-renderer": "^2.3.1",
@@ -94,5 +97,8 @@
     "last 1 chrome version",
     "last 1 firefox version",
     "last 1 safari version"
-  ]
+  ],
+  "jest": {
+    "testEnvironment": "jsdom"
+  }
 }
