+++ src/index.tsx
@@ -1,17 +1,4 @@
-import { useEffect, useState } from 'react';
-import Calendar, { Skeleton, type Props } from './component/ActivityCalendar';
-
-function ActivityCalendar(props: Props) {
-  // This component relies on various client hooks,
-  // so it cannot be rendered on the server.
-  const [isClient, setClient] = useState(false);
-
-  useEffect(() => {
-    setClient(true);
-  }, []);
-
-  return isClient ? <Calendar {...props} /> : null;
-}
+import ActivityCalendar, { Skeleton } from './component/ActivityCalendar';
 
 export { Skeleton };
 export default ActivityCalendar;
